add dev environment configuration
