+++ src/main/java/com/codurance/retropolis/config/Environment.java
@@ -3,5 +3,6 @@ package com.codurance.retropolis.config;
 public class Environment {
 
   public static final String PROD = "prod";
+  public static final String DEV = "dev";
   public static final String TEST = "test";
 }
