+++ src/main/java/com/codurance/retropolis/config/WebConfig.java
@@ -1,7 +1,14 @@
 package com.codurance.retropolis.config;
 
+import java.util.Collections;
+import org.springframework.boot.web.servlet.FilterRegistrationBean;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.core.Ordered;
+import org.springframework.web.cors.CorsConfiguration;
+import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
+import org.springframework.web.filter.CorsFilter;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
@@ -17,6 +24,20 @@ public class WebConfig {
     @Override
     public void addInterceptors(InterceptorRegistry registry) {
       registry.addInterceptor(new GoogleTokenAuthenticator());
+    }
+
+    @Bean
+    public FilterRegistrationBean<CorsFilter> simpleCorsFilter() {
+      UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
+      CorsConfiguration config = new CorsConfiguration();
+      config.setAllowCredentials(true);
+      config.setAllowedOrigins(Collections.singletonList("http://retropolis-fe.s3-website.eu-west-2.amazonaws.com"));
+      config.setAllowedMethods(Collections.singletonList("*"));
+      config.setAllowedHeaders(Collections.singletonList("*"));
+      source.registerCorsConfiguration("/**", config);
+      FilterRegistrationBean<CorsFilter> bean = new FilterRegistrationBean<>(new CorsFilter(source));
+      bean.setOrder(Ordered.HIGHEST_PRECEDENCE);
+      return bean;
     }
   }
 }
