+++ src/main/java/com/codurance/retropolis/config/WebDevConfig.java
@@ -1,0 +1,25 @@
+package com.codurance.retropolis.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@EnableWebMvc
+@Profile(Environment.DEV)
+public class WebDevConfig {
+
+  @Configuration
+  public class WebConfiguration implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+      registry.addMapping("/**")
+          .allowedMethods("*")
+          .allowedOrigins("http://localhost:3000");
+    }
+  }
+
+}
